requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
python-binance>=1.0.16

# Technical Analysis
#ta-lib>=0.4.24  # Richiede installazione separata del C++ TA-Lib
pandas-ta>=0.3.14b

# Machine Learning & Optimization
deap>=1.3.1
torch>=1.9.0
scikit-learn>=0.24.2

# Configuration & Logging
pyyaml>=5.4.1
python-dotenv>=0.19.0
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1

# Development Tools
black>=21.7b0  # Code formatter
flake8>=3.9.2  # Linter
mypy>=0.910    # Type checking
isort>=5.9.3   # Import sorting

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2

# Jupyter (for analysis and development)
jupyter>=1.0.0
ipykernel>=6.0.3

# Performance & Optimization
numba>=0.54.0
joblib>=1.0.1

# Data Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.1.0

# Database (optional, per future estensioni)
sqlalchemy>=1.4.23
pymongo>=3.12.0

# API & Web (optional, per future estensioni)
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0

# Dev Dependencies (optional)
pre-commit>=2.15.0      # Git hooks
jupyterlab>=3.1.0       # Enhanced Jupyter interface
notebook>=6.4.3         # Classic Jupyter notebook